+++ logparser/logparser.py
@@ -9,6 +9,7 @@ import os
 import platform
 import re
 import sys
+from telnetlib import DO, DONT, IAC, SB, SE, Telnet, TTYPE, WILL, WONT
 import time
 import traceback
 
@@ -16,21 +17,21 @@ try:
     from psutil import pid_exists
 except ImportError:
     pid_exists = None
+try:
+    from scrapy import __version__ as scrapy_version
+except ImportError:
+    scrapy_version = '1.0.0'
 
+from . import SETTINGS_PY_PATH
 from .__version__ import __version__
-from .constant import Constant
+from .common import Common
 from .scrapylogparser import parse
+from .utils import custom_settings, get_logger
 
 
-logger = logging.getLogger(__name__)
-_handler = logging.StreamHandler()
-# _formatter = logging.Formatter('%(asctime)s %(name)-12s %(levelname)-8s %(message)s')
-_formatter = logging.Formatter(fmt="[%(asctime)s] %(levelname)s in %(name)s: %(message)s",
-                               datefmt='%Y-%m-%d %H:%M:%S')
-_handler.setFormatter(_formatter)
-logger.addHandler(_handler)
+logger = get_logger(__name__)
 
-
+SUPPORTED_SCRAPY_VERSION = '1.5.1'
 IN_WINDOWS = True if platform.system() == 'Windows' else False
 SIMPLIFIED_KEYS = [
     'log_path',
@@ -45,24 +46,34 @@ SIMPLIFIED_KEYS = [
     'items',
     'first_log_time',
     'latest_log_time',
-    'elapsed',
+    'runtime',
     'shutdown_reason',
     'finish_reason',
     'last_update_time'
 ]
 
 
-class LogParser(Constant):
+# noinspection PyBroadException
+class LogParser(Common):
     # datas = {}  # Cause shared self.datas between test functions!
     logger = logger
 
-    def __init__(self, scrapyd_server, scrapyd_logs_dir, parse_round_interval, log_encoding,
-                 log_extensions, log_head_lines, log_tail_lines, delete_existing_json_files_at_startup,
-                 keep_data_in_memory, jobs_to_keep, chunk_size, verbose,
+    def __init__(self, scrapyd_server, scrapyd_logs_dir, parse_round_interval, enable_telnet,
+                 override_telnet_console_host, log_encoding, log_extensions, log_head_lines, log_tail_lines,
+                 delete_existing_json_files_at_startup, keep_data_in_memory, jobs_to_keep, chunk_size, verbose,
                  main_pid=0, debug=False, exit_timeout=0):
         self.SCRAPYD_SERVER = scrapyd_server
         self.SCRAPYD_LOGS_DIR = scrapyd_logs_dir
         self.PARSE_ROUND_INTERVAL = parse_round_interval
+        if enable_telnet and scrapy_version > SUPPORTED_SCRAPY_VERSION:
+            self.ENABLE_TELNET = False
+            self.logger.warning("ENABLE_TELNET is set to False for unsupported version of Scrapy: %s > %s",
+                                scrapy_version, SUPPORTED_SCRAPY_VERSION)
+        else:
+            self.ENABLE_TELNET = enable_telnet
+        self.tn = None
+
+        self.OVERRIDE_TELNET_CONSOLE_HOST = override_telnet_console_host
         self.LOG_ENCODING = log_encoding
         self.LOG_EXTENSIONS = log_extensions
         self.LOG_HEAD_LINES = log_head_lines
@@ -72,7 +83,8 @@ class LogParser(Constant):
         self.JOBS_TO_KEEP = jobs_to_keep
         self.CHUNK_SIZE = chunk_size
 
-        if verbose:
+        self.verbose = verbose
+        if self.verbose:
             self.logger.setLevel(logging.DEBUG)
         else:
             self.logger.setLevel(logging.WARNING)
@@ -98,7 +110,7 @@ class LogParser(Constant):
         if not os.path.exists(os.path.join(self.SCRAPYD_LOGS_DIR, 'stats.json')):
             self.save_text_into_logs_dir('stats.json', self.json_dumps(self.get_default_stats()))
 
-    def calc_elapsed(self, start_string, end_string):
+    def calc_runtime(self, start_string, end_string):
         try:
             start_datetime = datetime.strptime(start_string, '%Y-%m-%d %H:%M:%S')
             end_datetime = datetime.strptime(end_string, '%Y-%m-%d %H:%M:%S')
@@ -113,7 +125,7 @@ class LogParser(Constant):
         exit_condition_2 = not IN_WINDOWS and not self.check_pid(self.main_pid)
         if exit_condition_1 or exit_condition_2:
             sys.exit("!!! LogParser subprocess (pid: %s) exits "
-                     "since main_pid %s NOT exists" % (self.logparser_pid, self.main_pid))
+                     "since main_pid %s not exists" % (self.logparser_pid, self.main_pid))
 
     @staticmethod
     def check_pid(pid):
@@ -152,7 +164,7 @@ class LogParser(Constant):
         lines = re.split(r'\n', text)  # KEEP the same '\n'
         m = re.search(self.PATTERN_LOG_ENDING, text)
         if m:
-            self.logger.warning("Found log ending:%s", self.format_log_block('log ending', m.group()))
+            self.logger.info("Found log ending:%s", self.format_log_block('log ending', m.group()))
             text_to_ignore = ''
         else:
             # To ensure the integrity of a log with multilines, e.g. error with traceback info,
@@ -188,12 +200,13 @@ class LogParser(Constant):
     def get_default_stats(self):
         last_update_timestamp = int(time.time())
         last_update_time = datetime.fromtimestamp(last_update_timestamp).strftime('%Y-%m-%d %H:%M:%S')
-        return dict(status='ok', datas={}, logparser_version=self.logparser_version,
-                    last_update_timestamp=last_update_timestamp, last_update_time=last_update_time)
+        return dict(status='ok', datas={}, settings_py=self.handle_slash(SETTINGS_PY_PATH), settings=custom_settings,
+                    last_update_timestamp=last_update_timestamp, last_update_time=last_update_time,
+                    logparser_version=self.logparser_version)
 
     def handle_logfile(self, log_path):
         self.logger.debug("log_path: %s", log_path)
-        project, spider, job = self.parse_log_path(log_path)
+        project, spider, job, ext = self.parse_log_path(log_path)
         self.existing_file_keys.add(log_path)
 
         # The last modification time of Scrapy log does not change over time?!
@@ -205,6 +218,7 @@ class LogParser(Constant):
             self.logger.warning("New logfile found: %s (%s bytes)", log_path, size)
             json_path = os.path.join(self.SCRAPYD_LOGS_DIR, project, spider, '%s.json' % job)
             json_url = 'http://%s/logs/%s/%s/%s.json' % (self.SCRAPYD_SERVER, project, spider, job)
+            # NOTE: do not use handle_slash() on log_path, since parse_log_path() use os.sep
             data = OrderedDict(log_path=log_path, json_path=json_path, json_url=json_url,
                                size=size, position=0, status='ok', _head='')
             self.datas[log_path] = data
@@ -212,16 +226,21 @@ class LogParser(Constant):
             loaded_data = self.read_data(json_path)
             if loaded_data.get('size', -1) == size:
                 data.update(loaded_data)  # AVOID using data =
-                self.logger.warning("Skip parsing since data with same size found: %s (%s bytes)",
-                                    log_path, loaded_data['size'])
+                self.logger.warning("New logfile and its data with same size found: %s (size: %s) -> skip",
+                                    json_path, loaded_data['size'])
                 return
             else:
-                self.logger.debug("Go on to parse the new logfile.")
-        elif self.datas[log_path]['size'] == size:
-            self.logger.debug("Old logfile with old size: %s (%s bytes)", log_path, size)
+                self.logger.warning("New logfile: %s (%s bytes) -> parse", log_path, size)
+        elif size < self.datas[log_path]['size']:
+            self.logger.warning("Old logfile with smaller size: %s (before: %s, now: %s bytes) -> parse in next round",
+                                log_path, self.datas[log_path]['size'], size)
+            self.datas.pop(log_path)
+            return
+        elif size == self.datas[log_path]['size']:
+            self.logger.debug("Old logfile with old size: %s (%s bytes) -> skip", log_path, size)
             return
         else:
-            self.logger.warning("Old logfile with new size: %s (%s bytes)", log_path, size)
+            self.logger.warning("Old logfile with new size: %s (%s bytes) -> parse", log_path, size)
             data = self.datas[log_path]
 
             if not self.KEEP_DATA_IN_MEMORY:
@@ -245,7 +264,31 @@ class LogParser(Constant):
             if self.DEBUG:
                 self.save_text_into_logs_dir('appended_log.log', appended_log)
             self.parse_appended_log(data, appended_log)
+
         return data
+
+    @staticmethod
+    def handle_slash(string):
+        if not string:
+            return string
+        else:
+            return string.replace('\\', '/')
+
+    def handle_telnet(self, data):
+        data.setdefault('crawler_engine', {})
+        if (self.ENABLE_TELNET
+           and data['latest_matches']['telnet_console']
+           and not data['crawler_stats']):  # Do not telnet when the job is finished
+            try:
+                self.telnet(data)
+            except:
+                self.logger.error(traceback.format_exc())
+            finally:
+                if self.tn is not None:
+                    try:
+                        self.tn.close()
+                    except:
+                        pass
 
     @staticmethod
     def json_dumps(obj, sort_keys=False):
@@ -259,7 +302,7 @@ class LogParser(Constant):
             try:
                 self.run()
                 end_time = time.time()
-                self.logger.debug("Cost %.1f seconds this round", (end_time - start_time))
+                self.logger.debug("Took %.1f seconds this round", (end_time - start_time))
                 if 0 < self.EXIT_TIMEOUT < end_time - self.init_time:
                     self.logger.critical("GoodBye, EXIT_TIMEOUT: %s", self.EXIT_TIMEOUT)
                     break
@@ -268,20 +311,25 @@ class LogParser(Constant):
                     time.sleep(self.PARSE_ROUND_INTERVAL)
             except KeyboardInterrupt:
                 if self.main_pid:
-                    msg = "!!! LogParser subprocess (pid: %s) cancelled by KeyboardInterrupt" % self.logparser_pid
+                    self.logger.warning("LogParser subprocess (pid: %s) cancelled by KeyboardInterrupt",
+                                        self.logparser_pid)
                 else:
-                    msg = "!!! KeyboardInterrupt"
-                sys.exit(msg)
-            except Exception:
+                    self.logger.warning("KeyboardInterrupt")
+                sys.exit()
+            except:
                 self.logger.error(traceback.format_exc())
 
     def parse_appended_log(self, data, appended_log):
-        tail_last_time = data.get('tail', '')
+        tail_backup = data.get('tail', '')
         # Note that appended_log may be an empty string
         data_ = parse(appended_log, self.LOG_HEAD_LINES, self.LOG_TAIL_LINES)
         self.logger.debug("Parsed data_ from appended_log:\n%s", self.json_dumps(data_))
 
-        if data.get('first_log_time'):
+        if 'first_log_time' not in data:
+            # To keep the order of keys in Python 2
+            for k, v in data_.items():
+                data[k] = v
+        else:
             # data['head'] would be updated below
             data['tail'] = data_['tail']
 
@@ -291,11 +339,14 @@ class LogParser(Constant):
             if data_['latest_log_time'] != self.NA:
                 data['latest_log_time'] = data_['latest_log_time']
                 data['latest_log_timestamp'] = data_['latest_log_timestamp']
-            data['elapsed'] = self.calc_elapsed(data['first_log_time'], data['latest_log_time'])
+            data['runtime'] = self.calc_runtime(data['first_log_time'], data['latest_log_time'])
 
             data['datas'].extend(data_['datas'])
-            data['pages'] = data_['pages'] or data['pages']
-            data['items'] = data_['items'] or data['items']
+            for k in ['pages', 'items']:
+                if data[k] is None:
+                    data[k] = data_[k]
+                elif data_[k] is not None:
+                    data[k] = max(data[k], data_[k])
 
             for k, v in data_['latest_matches'].items():
                 data['latest_matches'][k] = v or data['latest_matches'][k]
@@ -304,6 +355,7 @@ class LogParser(Constant):
                 if data_['latest_matches'][k]:
                     data['%s_timestamp' % k] = data_['%s_timestamp' % k]
 
+            # "log_categories": {"critical_logs": {"count": 0, "details": []}}
             for k, v in data_['log_categories'].items():
                 if v['count'] > 0:
                     if data_['finish_reason'] != self.NA:
@@ -315,11 +367,9 @@ class LogParser(Constant):
             for k in ['shutdown_reason', 'finish_reason']:
                 if data_[k] != self.NA:
                     data[k] = data_[k]
-
+            data['crawler_stats'] = data_['crawler_stats'] or data['crawler_stats']
             data['last_update_timestamp'] = data_['last_update_timestamp']
             data['last_update_time'] = data_['last_update_time']
-        else:
-            data.update(data_)
 
         # To ensure the actual length of headlines and taillines
         if data['_head'] != self.LOG_HEAD_LINES:
@@ -337,11 +387,11 @@ class LogParser(Constant):
                 data['_head'] = self.LOG_HEAD_LINES
 
         if self.count_actual_lines(data['tail']) < self.LOG_TAIL_LINES:
-            if tail_last_time:
+            if tail_backup:
                 if appended_log:
-                    data['tail'] = '%s\n%s' % (tail_last_time, appended_log)
+                    data['tail'] = '%s\n%s' % (tail_backup, appended_log)
                 else:
-                    data['tail'] = tail_last_time
+                    data['tail'] = tail_backup
             else:
                 data['tail'] = appended_log
         data['tail'] = self.cut_text(data['tail'], self.LOG_TAIL_LINES, keep_head=False)
@@ -351,8 +401,8 @@ class LogParser(Constant):
     @staticmethod
     def parse_log_path(log_path):
         project, spider, _job = log_path.split(os.sep)[-3:]
-        job, __ = os.path.splitext(_job)  # ('job', '.log')
-        return project, spider, job
+        job, ext = os.path.splitext(_job)  # ('job', '') or ('job', '.log')
+        return project, spider, job, ext
 
     def read_appended_log(self, data, size=-1, backoff_times=10):
         # If the argument size is omitted, None, or negative, reads and returns all data until EOF.
@@ -398,13 +448,12 @@ class LogParser(Constant):
         data = {}
         self.logger.info("Try to load json file: %s", json_path)
         if not os.path.exists(json_path):
-            self.logger.warning("NOT found: %s", json_path)
+            self.logger.warning("Json file not found: %s", json_path)
         else:
             try:
                 with io.open(json_path, 'r', encoding='utf-8') as f:
                     data = json.loads(f.read())
             except Exception as err:
-                self.logger.warning("Fail to load json file: %s", json_path)
                 self.logger.error(err)
             else:
                 self.logger.debug("Loaded json file: %s", json_path)
@@ -424,10 +473,12 @@ class LogParser(Constant):
         for log_path in self.log_paths:
             try:
                 data = self.handle_logfile(log_path)
-                if data:
-                    self.save_data(data)
-            except Exception as err:
-                self.logger.error(err)
+                if not data:
+                    continue
+                self.handle_telnet(data)
+                self.save_data(data)
+            except:
+                self.logger.error(traceback.format_exc())
                 self.logger.warning("Pop %s from self.datas", log_path)
                 self.datas.pop(log_path, None)
 
@@ -439,7 +490,6 @@ class LogParser(Constant):
         self.save_datas()
 
     def save_data(self, data):
-        data['logparser_version'] = self.logparser_version
         with io.open(data['json_path'], 'wb') as f:
             f.write(self.json_dumps(data).encode('utf-8', 'replace'))
         self.logger.warning("Saved to %s", data['json_path'])
@@ -449,7 +499,10 @@ class LogParser(Constant):
         for log_path, data in self.datas.items():
             if self.KEEP_DATA_IN_MEMORY and log_path in self.existing_file_keys:
                 data = self.simplify_data(dict(data))
-            project, spider, job = self.parse_log_path(log_path)
+            else:
+                data = dict(data)
+            data.pop('_head')  # To simplify data for 'List Stats' in the Overview page
+            project, spider, job, ext = self.parse_log_path(log_path)
             stats['datas'].setdefault(project, {})
             stats['datas'][project].setdefault(spider, {})
             stats['datas'][project][spider][job] = data
@@ -501,3 +554,102 @@ class LogParser(Constant):
             self.logger.info("Now all_keys in memory: %s", len(self.datas))
         else:
             self.logger.info("all_keys in memory: %s", len(self.datas))
+
+    # Note that the telnet feature is tested in test_telnet_in_stats() by ScrapydWeb
+    # https://stackoverflow.com/questions/18547412/python-telnetlib-to-connect-to-scrapy-telnet-to-read-stats
+    def telnet(self, data):
+        # Telnet console listening on 127.0.0.1:6023
+        m = re.search(r'listening[ ]on[ ](.+):(\d+)$', data['latest_matches']['telnet_console'])
+        if not m:
+            self.logger.warning("Fail to extract host and port from %s", data['latest_matches']['telnet_console'])
+            return
+        host, port = m.groups()
+        host = self.OVERRIDE_TELNET_CONSOLE_HOST or host
+        self.logger.debug("Try to telnet to %s:%s for %s", host, port, data['log_path'])
+        try:
+            self.tn = Telnet(host, int(port), timeout=10)
+        # except ConnectionRefusedError as err:  # Python 2: <class 'socket.error'>
+        except Exception as err:
+            self.logger.error("Fail to telnet to %s:%s for %s: %s", host, port, data['log_path'], err)
+            return
+
+        if self.verbose:
+            self.tn.set_debuglevel(logging.DEBUG)
+        # [twisted] CRITICAL: Unhandled Error
+        # Failure: twisted.conch.telnet.OptionRefused: twisted.conch.telnet.OptionRefused
+        # https://github.com/jookies/jasmin-web/issues/2
+        self.tn.set_option_negotiation_callback(self.telnet_callback)
+
+        # spider._job, spider._version, settings.attributes["BOT_NAME"].value, JOB, SPIDER, PROJECT
+        # '\'logs\\\\demo_persistent\\\\test\\\\2019-01-23T18_25_34.log\'\r\r\r\n>>>'
+        _log_file = self.telnet_io(b'settings.attributes["LOG_FILE"].value', return_text=True)
+        self.logger.debug("settings['LOG_FILE'] found via telnet: %s", _log_file)
+        # Username: Password:
+        if 'Username:' in _log_file:
+            self.logger.error("The feature of collecting crawler_stats and crawler_engine via telnet "
+                              "temporarily only works for Scrapy 1.5.1 and its earlier version: %s", _log_file)
+            return
+        for part in self.parse_log_path(data['log_path']):
+            if part not in _log_file:
+                self.logger.warning("%s not found in settings['LOG_FILE']: %s", part, _log_file)
+                return
+
+        data['crawler_stats'] = self.telnet_io(b'stats.get_stats()') or data['crawler_stats']
+        if data['crawler_stats']:
+            # update_log_count=False to avoid wrong count in parse_appended_log() when the job is running
+            self.update_data_with_crawler_stats(data, data['crawler_stats'], update_log_count=False)
+        data['crawler_engine'] = self.telnet_io(b'est()') or data['crawler_engine']
+
+        self.logger.debug("crawler_stats:\n%s", self.json_dumps(data['crawler_stats']))
+        self.logger.debug("crawler_engine:\n%s", self.json_dumps(data['crawler_engine']))
+
+    @staticmethod
+    def telnet_callback(tn, command, option):
+        if command == DO and option == TTYPE:
+            tn.sendall(IAC + WILL + TTYPE)
+            tn.sendall(IAC + SB + TTYPE + '\0' + 'LogParser' + IAC + SE)
+        elif command in (DO, DONT):
+            tn.sendall(IAC + WILL + option)
+        elif command in (WILL, WONT):
+            tn.sendall(IAC + DO + option)
+
+    def telnet_io(self, command, return_text=False):
+        # Microsoft Telnet> o
+        # ( to )127.0.0.1 6023
+        # >>>stats.get_stats()
+        # >>>est()
+        self.tn.write(b'%s\n' % command)
+        content = self.tn.read_until(b'\n>>>', timeout=10)
+        # print(repr(content))
+        # b"\x1bc>>> \x1b[4hstats.get_stats()\r\r\r\n{'log_count/INFO': 61,
+        # 'start_time': datetime.datetime(2019, 1, 22, 9, 7, 14, 998126),
+        # 'httperror/response_ignored_status_count/404': 1}\r\r\r\n>>>"
+        # b' est()\r\r\r\nExecution engine status\r\r\r\n\r\r\r\n
+        # time()-engine.start_time                        : 3249.7548048496246
+        # engine.scraper.slot.needs_backout()             : False\r\r\r\n\r\r\r\n\r\r\r\n>>>'
+        text = content.decode('utf-8')
+        if return_text:
+            return text
+        m = re.search(r'{.+}', text)
+        if m:
+            result = self.parse_crawler_stats(m.group())
+        else:
+            lines = [line for line in re.split(r'\r\n|\n|\r', text) if ':' in line]
+            result = dict([re.split(r'\s*:\s*', line, maxsplit=1) for line in lines])
+            for k, v in result.items():
+                if k == 'engine.spider.name':
+                    continue
+                elif v == 'True':
+                    result[k] = True
+                elif v == 'False':
+                    result[k] = False
+                else:
+                    try:
+                        result[k] = int(float(v))
+                    except (TypeError, ValueError):
+                        pass
+
+        if result:
+            return self.get_ordered_dict(result, source='telnet')
+        else:
+            return {}