+++ logparser/scrapylogparser.py
@@ -4,7 +4,8 @@ from datetime import datetime
 import re
 import time
 
-from .constant import Constant
+from .__version__ import __version__
+from .common import Common
 
 
 def parse(text, headlines=100, taillines=200):
@@ -24,7 +25,8 @@ def parse(text, headlines=100, taillines=200):
     return ScrapyLogParser(text, headlines, taillines).main()
 
 
-class ScrapyLogParser(Constant):
+# noinspection PyTypeChecker
+class ScrapyLogParser(Common):
 
     def __init__(self, text, headlines=100, taillines=200):
         text = text.strip()
@@ -33,7 +35,7 @@ class ScrapyLogParser(Constant):
         self.data['head'] = '\n'.join(self.lines[:headlines])
         self.data['tail'] = '\n'.join(self.lines[-taillines:])
         # Modify text for self.DATAS_PATTERN, self.LOG_CATEGORIES_PATTERN_DICT
-        self.text = '\n%s\n2019-01-01 00:00:01 DEBUG' % text
+        self.text = '\n%s\n2019-01-01 00:00:01 [] DEBUG' % text
 
     def main(self):
         self.extract_time()
@@ -42,8 +44,8 @@ class ScrapyLogParser(Constant):
         self.extract_log_categories()
         self.extract_shutdown_reason()
         self.extract_stats_dumped()
-        self.data['last_update_timestamp'] = int(time.time())
-        self.data['last_update_time'] = self.timestamp_to_string(self.data['last_update_timestamp'])
+        self.data['last_update_time'], self.data['last_update_timestamp'] = self.get_current_time_timestamp()
+        self.data['logparser_version'] = __version__
         return self.data
 
     def re_search_final_match(self, pattern, default='', step=-1):
@@ -58,11 +60,11 @@ class ScrapyLogParser(Constant):
 
     @staticmethod
     def datetime_obj_to_timestamp(datetime_obj):
+        """
+        :param datetime_obj: datetime.datetime
+        :rtype: int object
+        """
         return int(time.mktime(datetime_obj.timetuple()))
-
-    @staticmethod
-    def timestamp_to_string(timestamp):
-        return datetime.fromtimestamp(timestamp).strftime('%Y-%m-%d %H:%M:%S')
 
     def extract_time(self):
         self.data['first_log_time'] = self.re_search_final_match(r'^%s[ ]' % self.DATETIME_PATTERN, step=1)[:19]
@@ -71,13 +73,13 @@ class ScrapyLogParser(Constant):
         if self.data['first_log_time'] and self.data['latest_log_time']:
             first_log_datetime = self.string_to_datetime_obj(self.data['first_log_time'])
             latest_log_datetime = self.string_to_datetime_obj(self.data['latest_log_time'])
-            self.data['elapsed'] = str(latest_log_datetime - first_log_datetime)
+            self.data['runtime'] = str(latest_log_datetime - first_log_datetime)
             self.data['first_log_timestamp'] = self.datetime_obj_to_timestamp(first_log_datetime)
             self.data['latest_log_timestamp'] = self.datetime_obj_to_timestamp(latest_log_datetime)
         else:
             self.data['first_log_time'] = self.NA
             self.data['latest_log_time'] = self.NA
-            self.data['elapsed'] = self.NA
+            self.data['runtime'] = self.NA
             self.data['first_log_timestamp'] = 0
             self.data['latest_log_timestamp'] = 0
 
@@ -88,13 +90,24 @@ class ScrapyLogParser(Constant):
         self.data['datas'] = [[str(time_), int(pages), int(pages_min), int(items), int(items_min)]
                               for (time_, pages, pages_min, items, items_min) in datas]
         # TODO: Crawled (200) <GET,     Scraped from <200
-        self.data['pages'] = int(datas[-1][1]) if datas else 0
-        self.data['items'] = int(datas[-1][3]) if datas else 0
+        self.data['pages'] = None
+        self.data['items'] = None
+        if self.data['datas']:
+            self.data['pages'] = max(i[1] for i in self.data['datas'])
+            self.data['items'] = max(i[3] for i in self.data['datas'])
 
     def extract_latest_matches(self):
         self.data['latest_matches'] = {}
         for k, v in self.LATEST_MATCHES_PATTERN_DICT.items():
-            self.data['latest_matches'][k] = self.re_search_final_match(v, step=1 if k == 'resuming_crawl' else -1)
+            step = 1 if k in ['telnet_console', 'resuming_crawl'] else -1
+            self.data['latest_matches'][k] = self.re_search_final_match(v, step=step)
+        # Scrapyd in PY2: u"{u'Chinese \\u6c49\\u5b57 1':"
+        latest_item = self.data['latest_matches']['latest_item']
+        if '\\u' in latest_item:  # and sys.version_info.major < 3:
+            try:
+                self.data['latest_matches']['latest_item'] = latest_item.encode('utf-8').decode('unicode-escape')
+            except:
+                pass
         # latest_crawl_timestamp, latest_scrape_timestamp
         for k in ['latest_crawl', 'latest_scrape']:
             time_string = self.data['latest_matches'][k][:19]
@@ -120,32 +133,15 @@ class ScrapyLogParser(Constant):
         self.data['shutdown_reason'] = m.group(1) if m else self.NA
 
     def extract_stats_dumped(self):
-        # TODO
-        # 'finish_time': datetime.datetime(2018, 10, 23, 10, 29, 41, 174719),
-        # 'start_time': datetime.datetime(2018, 10, 23, 10, 28, 35, 70938)}
-
-        # 'finish_reason': 'closespider_timeout',
-        m = re.search(r":[ ]'(.+?)'", self.re_search_final_match(r"'finish_reason'"))
-        self.data['finish_reason'] = m.group(1) if m else self.NA
-
-        if self.data['finish_reason'] != self.NA:
-            redirect_count = sum([int(i) for i in re.findall(self.RESPONSE_STATUS_REDIRECT_PATTERN, self.text)])
-            if redirect_count > 0:
-                self.data['log_categories']['redirect_logs']['count'] = redirect_count
-
-            # 'downloader/response_count': 4,
-            # 'downloader/response_status_count/200': 2,
-            # 'downloader/response_status_count/302': 1,
-            # 'downloader/response_status_count/404': 1,
-            # 'response_received_count': 3,
-            m = re.search(r":[ ](\d+)", self.re_search_final_match(r"'response_received_count'"))
-            self.data['pages'] = (int(m.group(1)) if m else 0) or self.data['pages']
-
-            # 'item_scraped_count': 2,
-            m = re.search(r":[ ](\d+)", self.re_search_final_match(r"'item_scraped_count'"))
-            self.data['items'] = (int(m.group(1)) if m else 0) or self.data['items']
-
-            for level, pattern in self.STATS_DUMPED_PATTERN_DICT.items():
-                m = re.search(r":[ ](\d+)", self.re_search_final_match(pattern))
-                if m:
-                    self.data['log_categories'][level]['count'] = int(m.group(1))
+        self.data['finish_reason'] = self.NA  # May be updated in update_data_with_crawler_stats()
+        m = re.search(self.PATTERN_LOG_ENDING, self.text)
+        if not (m and m.group(3)):
+            self.data['crawler_stats'] = {}
+        else:
+            crawler_stats = self.parse_crawler_stats(m.group(3))
+            self.update_data_with_crawler_stats(self.data, crawler_stats, update_log_count=True)
+            self.data['crawler_stats'] = self.get_ordered_dict(crawler_stats, source='log')
+            time_string = m.group(1)
+            datetime_obj = self.string_to_datetime_obj(time_string)
+            self.data['crawler_stats']['last_update_time'] = time_string
+            self.data['crawler_stats']['last_update_timestamp'] = self.datetime_obj_to_timestamp(datetime_obj)